+++ demo/src/test/java/com/carlos/cutils/demo/PhoneTest.java
@@ -2,6 +2,7 @@ package com.carlos.cutils.demo;
 
 import org.junit.Test;
 
+import java.io.File;
 import java.util.regex.Pattern;
 
 /**
@@ -12,8 +13,20 @@ public class PhoneTest {
 
     @Test
     public void test() {
-        String phone = "12334434343434434";
-        System.out.println(isPhoneNumber(phone));
+//        String phone = "12334434343434434";
+//        System.out.println(isPhoneNumber(phone));
+
+        int i =3;
+        System.out.println((i++)+(i++)+(++i));
+
+//        String a =new String("abc ");
+//        String b =a.trim();
+//        String c =b.intern();
+//        System.out.println(c==b);
+//
+//        File f = new File(this.getClass().getResource("/").getPath());
+//        System.out.println(f);
+//        getFiles(f.getPath());
     }
 
     public static boolean isPhoneNumber(String input) {// 判断手机号码是否规则
@@ -22,4 +35,17 @@ public class PhoneTest {
         return p.matches(regex, input);//如果不是号码，则返回false，是号码则返回true
 
     }
+
+    public void getFiles(String path) {
+        File file = new File(path);
+        if (!file.exists()) return;
+        for (String sfilePath : file.list()) {
+            System.out.println("file path:" + sfilePath);
+            File file1 = new File(sfilePath);
+            if (file1.exists())
+                getFiles(sfilePath);
+        }
+    }
+
+
 }