+++ demo/src/test/java/com/carlos/cutils/demo/algorithm/JosephusProblem.kt
@@ -1,0 +1,8 @@
+package com.carlos.cutils.demo.algorithm
+
+/**
+ * Created by Carlos on 2020/3/19.
+ */
+class JosephusProblem {
+
+}